requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
biopython==1.83
blinker==1.6.2
branca==0.6.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
Flask==3.0.0
folium==0.14.0
fonttools==4.47.2
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.4.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.0
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
stack-data==0.6.3
Tcl==0.2
tornado==6.3.3
traitlets==5.11.2
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
